requirements.txt
appdirs==1.4.4
asgiref==3.2.7
certifi==2020.4.5.2
cssselect==1.1.0
d==0.2.2
distlib==0.3.0
dj-database-url==0.5.0
Django==2.1.15
django-bootstrap-form==3.4
django-filter==2.2.0
django-heroku==0.3.1
django-json==0.27
django-messages==0.6.0
django-reset-migrations==0.4.0
django-widget-tweaks==1.4.8
et-xmlfile==1.0.1
filelock==3.0.12
gunicorn==20.0.4
jdcal==1.4.1
lxml==4.5.1
Markdown==3.2.2
mysqlclient==1.4.6
openpyxl==3.0.3
Pillow==7.1.2
pipenv==2020.6.2
psycopg2==2.8.5
psycopg2-binary==2.8.5
Pygments==2.6.1
PyMySQL==0.9.3
pyquery==1.4.1
pytz==2020.1
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
virtualenv==20.0.20
virtualenv-clone==0.5.4
whitenoise==5.1.0
